requirements.txt
absl-py==1.3.0
aggdraw==1.3.15
apache-beam==2.37.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
boto3==1.18.36
botocore==1.21.36
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
clang==14.0
cloudpickle==2.2.0
contourpy==1.0.5
crcmod==1.7
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.4
executing==1.1.1
fastavro==1.7.0
fasteners==0.18
fastjsonschema==2.16.2
flatbuffers==22.9.24
fonttools==4.38.0
gast==0.4.0
google-api-core==1.31.5
google-api-python-client==2.65.0
google-apitools==0.5.32
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.13.1
google-cloud-bigtable==1.7.1
google-cloud-core==1.7.2
google-cloud-datastore==1.15.4
google-cloud-dlp==3.6.2
google-cloud-language==1.3.1
google-cloud-pubsub==2.12.0
google-cloud-pubsublite==1.4.1
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.2
google-cloud-videointelligence==1.16.2
google-cloud-vision==1.0.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
h5py==3.7.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
importlib-resources==5.10.0
inquirer==2.10.0
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labelImg==1.8.5
libclang==14.0.6
lime==0.2.0.1
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.3.0
nest-asyncio==1.5.6
nets==0.0.3.1
networkx==2.8.7
nltk==3.2.4
notebook==6.4.10
numba==0.56.3
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.1
overrides==7.3.1
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
path==16.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
preprocessing==0.1.13
prometheus-client==0.15.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
pymongo==4.3.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.5
PyWavelets==1.4.1
pyzmq==24.0.1
readchar==4.0.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
shap==0.39.0
six==1.16.0
slicer==0.0.7
soupsieve==2.3.2.post1
sphinx-rtd-theme==0.2.4
stack-data==0.5.1
stem==1.8.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.18.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.10.0
tensorflow-model-analysis==0.38.0
tensorflow-serving-api==2.8.0
termcolor==2.0.1
terminado==0.17.0
tf-estimator-nightly==2.8.0.dev2021122109
tfx-bsl==1.7.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typeguard==2.13.3
typing-utils==0.1.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
visualkeras==0.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.10.0
